requirements.txt
accelerate==1.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
auraloss==0.4.0
blobfile==3.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
decorator==5.1.1
deepspeed==0.15.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
einx==0.3.0
exceptiongroup==1.2.2
filelock==3.16.1
fire==0.7.0
fonttools==4.54.1
frozendict==2.4.5
frozenlist==1.4.1
fsspec==2024.6.1
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
hjson==3.1.0
huggingface-hub==0.25.1
hydra-core==1.3.2
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.11.7
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
meson==1.4.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.1
pandas==2.2.3
pesq==0.0.4
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyparsing==3.1.4
pystoi==0.4.1
pytest==8.3.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rotary-embedding-torch==0.8.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentencepiece==0.2.0
sentry-sdk==2.14.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.3
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
tomli==2.0.2
torchao==0.5.0
torch==2.4.1
torchaudio==2.4.1
torchmetrics==1.4.2
torchtune==0.3.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vector-quantize-pytorch==1.17.8
wandb==0.18.1
xxhash==3.5.0
yarl==1.12.1
zipp==3.20.2
shutil
s3prl
pystoi
